add more debug

+++ ssh.sh
@@ -6,7 +6,7 @@
 # create new agent, add id
 TMPDIR=${TMPDIR:-/tmp}
 SSH_DEBUG=${SSH_DEBUG:-0}
-ssh_debug() { [[ ${SSH_DEBUG} == "0" ]] || echo "$@"; }
+ssh_debug() { [[ ${SSH_DEBUG} == "0" ]] || >&2 echo "$@"; }
 
 ssh_config_get_key_by_host() {
 	local host="$1" host_re="unmatchable" id=""
@@ -15,37 +15,41 @@ ssh_config_get_key_by_host() {
 	cat ${HOME}/.ssh/config | while read key value; do
 		[[ "$key" == "Host" ]] && { host_re="$value"; id="$default_id"; }
 		[[ "$key" == "IdentityFile" ]] && id="$value"
-		[[ "$key" == "" && $host = $host_re ]] && { echo $id; break; }
+		[[ "$key" == "" && $host =~ $host_re ]] && { echo $id; break; }
 	done
 }
 ssh_agent_list() {
-	find ${HOME}/.ssh/agent ${HOME}/.ssh-agent* {$TMPDIR,/tmp}/ssh-*/agent* -user $USER -type s 2>/dev/null
+	find ${HOME}/.ssh/agent ${HOME}/.ssh-agent* {$TMPDIR,/tmp}/ssh-*/agent* -user $USER -type s 2>/dev/null | sort | uniq
 }
 ssh_find_sock() {
 	local key="$1"
-	ssh_agent_list | while read sock; do
+	for sock in $( ssh_agent_list ); do
 		SSH_AUTH_SOCK=$sock ssh-add -l 2>&1 | grep -qs "$key" && { echo $sock; break; }
 	done
 }
 ssh_agent_check() {
 	local file=$(ssh_config_get_key_by_host $1 | sed -e "s#~#${HOME}#" )
 	local key="."
-	ssh_debug "agent check got file $file"
+	ssh_debug "agent check got file $file for host $1"
 	[[ -f "$file"  ]] && key=$( ssh-keygen -l -f $file | awk '{print $2}' )
 	ssh_debug "looking for id $key"
 	local sock=$( ssh_find_sock $key  )
-	ssh_debug "for this $key got auth sock $sock"
+	ssh_debug "for this $key got auth sock '$sock'"
 	if [[ -e $sock ]]; then
 		export SSH_AUTH_SOCK=$sock
 		return 0
 	fi
 	ssh_debug "no agent found, creating new "
 	eval `ssh-agent` &>/dev/null
-	ssh-add $file &>/dev/null
+	ssh_debug "new agent in ${SSH_AUTH_SOCK}"
+	ssh_debug "ssh-add $file"
+	ssh-add $file
+	ssh_debug $( ssh-add -l )
 }
 ssh_agent_clean() {
 	local a
 	for a in /tmp/ssh*/agent*; do
+		[[ -e $a ]] || continue
 		SSH_AUTH_SOCK=$a ssh-add -l && continue
 		mv -v $( dirname $a ) /tmp/foo-$( basename $a )
 	done
